Fix (MSVC) compiler warnings
reg. size_t vs. int, after #11194

+++ src/engraving/libmscore/scoretree.cpp
@@ -874,7 +874,7 @@ EngravingObject* BSymbol::scanChild(int idx) const
 
 int BSymbol::scanChildCount() const
 {
-    return _leafs.size();
+    return static_cast<int>(_leafs.size());
 }
 
 //---------------------------------------------------------
